requirements.txt
Flask~=3.1.2
python-dotenv~=1.2.1
password-validator==1.0
wtforms==3.2.1
Flask-WTF==1.2.2